move : stage.handler.js
위치 변경
